ex 4

+++ Lista 01/ex_03.java
@@ -1,11 +1,6 @@
 import java.util.Scanner;
 
 public class ex_03 {
-  /**
-   * ax*x + bx + c
-   * Delta = b**2 -4 * a * c
-   * (-b +- sqrt(Delta))/2a
-   */
   public static void main(String[] args) {
     Scanner keyboard = new Scanner(System.in);
     double salario, salarioFinal, imposto, gratificacao;
@@ -15,7 +10,7 @@ public class ex_03 {
     gratificacao = 5;
     
     System.out.println("Digite o seu salário:");
-    salario = keyboard.nextFloat();
+    salario = keyboard.nextDouble();
     salarioFinal = (100+gratificacao-imposto)/100*salario;
 
     System.out.println("Salario Final:"+salarioFinal);
